chore: print sha256sum only in release script
When running the `sha256sum --binary` command with input from standard
in, the output is:

$SHA_HERE *-

As noted in the man page for sha256sum, the asterix indicates "binary
mode." The trailing dash means standard in.

This commit removes the extraneous information such that only the SHA
will be printed in the final output.

See https://unix.stackexchange.com/questions/487028/print-sha-sums-without-at-the-end
for more.